step 9 Added comments to increase documentation in our auth service files. Migrated our AuthService into our view files. Added an initialization function into our Auth files.
