+++ lib/services/auth/auth_provider.dart
@@ -1,7 +1,11 @@
 import 'package:mynotes/services/auth/auth_user.dart';
 
 abstract class AuthProvider {
-  AuthUser? get currentUser;
+  AuthUser? get currentUser; // Either return an AuthUser object or return null
+
+  // Call await on all methods that return a Future
+
+  Future<void> initialize();
 
   Future<AuthUser> logIn({
     required String email,
